backend/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
astroid==2.12.4
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
dill==0.3.5.1
distlib==0.3.6
ecdsa==0.18.0
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.3.1
pylint==2.15.0
python-dotenv==0.20.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
pytz==2022.2.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.40
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
